requirements.txt
torch
numpy
Pillow
flask
flask_cors
pymongo
torchvision
python-dateutil